feat: add endpoints to friends-request controller

+++ backend/src/friends/friend-requests.service.ts
@@ -1,7 +1,7 @@
 import { HttpException, Injectable } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import mongoose, { ClientSession, Connection, Model } from 'mongoose';
-import { Friends } from 'src/schemas/Friends-requests.schema';
+import { FriendRequests } from 'src/schemas/Friends-requests.schema';
 import { FriendsDocument } from 'src/utils/schema.types';
 import { FriendsFilterType } from './utils/types';
 import { UsersService } from 'src/users/users.service';
@@ -10,7 +10,8 @@ import { InjectConnection } from '@nestjs/mongoose';
 @Injectable()
 export class FriendsService {
   constructor(
-    @InjectModel(Friends.name) private friendsModel: Model<FriendsDocument>,
+    @InjectModel(FriendRequests.name)
+    private friendsModel: Model<FriendsDocument>,
     @InjectConnection() private readonly connection: Connection,
     private usersService: UsersService,
   ) {}
@@ -40,9 +41,7 @@ export class FriendsService {
   async sendRequest(targetId: string, userId: string) {
     const target = await this.usersService.findById(targetId);
     if (!target) throw new HttpException('User not found', 404);
-    const isFriends = target.friends?.includes(
-      new mongoose.Types.ObjectId(userId),
-    );
+    const isFriends = target.friends?.some((id) => id.toString() === userId);
     if (isFriends) throw new HttpException('Friendship already exist', 409);
     const pendingRequest = await this.getFriendship(targetId, userId);
     if (pendingRequest) throw new HttpException('Request already exist', 409);
@@ -54,23 +53,25 @@ export class FriendsService {
     const session = await this.connection.startSession();
     session.startTransaction();
 
-    try {
-      const request = await this.friendsModel
-        .findById(requestId)
-        .session(session);
-      if (!request) throw new HttpException('Request does not exist', 404);
-      if (request.to.toString() !== userId.toString())
-        throw new HttpException(
-          'You are not authorized to accept this request',
-          403,
-        );
-      const sender = await this.usersService.findById(request.from.toString());
-      if (!sender) {
-        await this.friendsModel.deleteOne({ _id: requestId });
-        throw new HttpException('User no longer exist.', 404);
-      }
+    const request = await this.friendsModel.findById(requestId);
 
-      await this.friendsModel.deleteOne({ _id: request._id });
+    if (!request) throw new HttpException('Request does not exist', 404);
+    if (request.to.toString() !== userId.toString())
+      throw new HttpException(
+        'You are not authorized to accept this request',
+        403,
+      );
+    if (request.from.toString() === userId) {
+      throw new HttpException('Request cannot accept your own request', 409);
+    }
+    const sender = await this.usersService.findById(request.from.toString());
+    if (!sender) {
+      await this.friendsModel.deleteOne({ _id: requestId });
+      throw new HttpException('User no longer exist.', 404);
+    }
+
+    try {
+      await this.friendsModel.deleteOne({ _id: request._id }).session(session);
 
       await Promise.all([
         this.usersService.addFriend(sender._id, userId, session),
@@ -92,10 +93,7 @@ export class FriendsService {
   }
 
   async declineFriendship(targetId: string, userId: string) {
-    const request = await this.friendsModel.findOne({
-      from: targetId,
-      to: userId,
-    });
+    const request = await this.getFriendship(targetId, userId);
     if (!request)
       throw new HttpException('Pending request friendship not found.', 404);
